+++ src/components/Results.tsx
