+++ src/components/ShabadBlock.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { FlatList } from 'react-native'
 
-import { Element } from '../../../types/types'
+import { Element } from '../../types/types'
 import {
   useSelectionInfo,
   useValues,
-} from '../../store/StateHooks'
+} from '../store/StateHooks'
 import LineBlock from './LineBlock'
 
 const ShabadBlock = ( {
